Fix renaming of IPSW and update readme
When the two buildID's were equal, "downloaded.ipsw" wasn't getting renamed to "current.ipsw", because I had put a $ symbol infront of "downloaded.ipsw", for some reason. That is fixed now. I also added to confirm that this works on OS X Yosemite, and also added that iTunes can show the idevice model as "Product Type".